requirements.txt
requests==2.31.0
pandas==1.3.5
MLB-StatsAPI==1.6.2